refactor(Models): Refactor models to attempt preventing direct usage of base models.
